Mount only relevant files into backend container
Directories like tmp and log should not be mounted into the backend
container. When for example the rails process does not terminate
correctly, the pid file in tmp might not be cleaned up. Resetting the
environment should fix this, but it does not, when the pid file is
copied or mounted into a new container.