Set up a staging environment
We had a serious outage last night when I attempted to update the
postgres version supplied by our hoster uberspace.de. A better solution
would have been a staging environment similar to the production
environment where we can safely try out the manual update of postgres.

This commit also includes a small fix, if you forget the trailing / at
the end of the deploy folder for the frontend, the release versions will
be uploaded to the parent folder, in our case $HOME.