Refactoring: Use serializers to define json output format
Broadcast#to_graph_node and Similarity#to_graph_edge where replaced with 3 serializers, NodeSerializer, LinkSerializer and GraphSerializer. The last one works as container for the first two.
This solution is not compatible with the current master, because it uses a different output format for similarities action, in particular it adds a root element 'data', which contains the 2 elements 'nodes' and 'links':

Example output:
{ 'data' => { 'nodes' => [], 'links' => [] } }