Remove login route altogether
If we have a login route, it will be added to our history. So, when the
authentication sends the user to the last route, this will always be
/login.

The reason for having the login route is to give the user an explanation
for authentication. It is questionable at best, if the user is even
reading the description.

Therefore existing and satisfying behaviour over speculative benefit