Create github templates
Added standard templates pulled from around the web.
Commented out the descriptions under the heading so that the user does
not need to delete the verbose text in the template.

This is being done to standardize bug reports, feature requests, and
pull requests so that we can have good documentation of the changes
being made. [ci skip]

Fixes #848