Cd back to the root directory to fix deployment
@faithngetich and @ciremoussadia I believe that our deployment errored
because it tried to decrypt a file that did not exist in the current
folder. If you look closely, the last command in `script` part of the
`.travis.yml` file changes into the backend directory. Apparently,
that's the directory where Travis tries to decrypt the SSH private key
for deployment. But there is not `deploy_rsa.enc` file in the backend
directory, it's in the root directory.

So, this is a great example that shows what can go possibly wrong when
changing the build configuration :wink: